Original data used by Grauvogel/von Soest (2014)

A dataset containing the calibrated set values for the article: Grauvogel, Julia and Christian von Soest (2014): Claims to Legitimacy Count: Why Sanctions Fail to Instigate Democratisation in Authoritarian Regimes. European Journal of Political Research 53 (4): 635-653.

Usage

Grauvogel2014

Format

A data frame with 120 rows and 10 variables:

Code

Sender-target ID

Sender

Country or institution imposing sanctions

Target

Country that is target of sanctions

Timeframe

Considered years for each country case

Persistence

Degree of regime persistence after the intervention

Comprehensiveness

Scope of the imposed sanctions - comprehensive vs. targeted sanctions

Linkage

Economic and social, respectively communicative and geographic ties

Vulnerability

Military and economic vulnerability of the state to outside pressure

Repression

Degree of repression by the state

Claims

Variety and strength of claims to legitimacy

Source

Grauvogel (2014) <doi:10.1111/1475-6765.12065>