Original data used by Thiem (2011)

A dataset containing the calibrated set values for the article: Thiem, Alrik (2011): Conditions of Intergovernmental Armaments Cooperation in Western Europe, 1996-2006. European Political Science Review 3 (1): 1-33.

Usage

Thiem2011

Format

A data frame with 165 rows and 10 variables:

id

Country-year ID

year

Time ID

country

Country ID

memberfs

Monadic count of membership in formal intergovernmental agreements on armaments cooperation

fedismfs

Degree to which a country’s domestic constitutional setup is federalist in character

homogtyfs

Bilateral interaction scores based on all UN and NATO military missions conducted between 1996 and 2006

powdifffs

Score to measure a country's military power based on the CINC score

comptvnsfs

Competitiveness of a country’s domestic armaments industry

pubsupfs

Public support for cooperation in defence

ecodpcefs

Degree of economic dependence

Source

Thiem(2011) <doi:10.1017/S1755773910000251>